+++ CHANGELOG.md
@@ -19,6 +19,7 @@ contents to a new release section (e.g. _[0.x]_).
 
 - Fixed typo in simple solution rendering (PLUT-568)
 - Fixed package CI job by setting JVM back to 17 (PLUT-570)
+- Set output of `EvaluateArithmeticExpression` plan to be an integer (PLUT-550)
 
 ### Changed